requirements.txt
argcomplete==1.12.3
attrs==21.4.0
Authlib==0.15.5
black==22.1.0
CacheControl==0.12.10
cached-property==1.5.2
cachelib==0.6.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
cryptography==36.0.1
distlib==0.3.4
filelock==3.4.2
firebase-admin==5.2.0
Flask==2.0.2
google-ads==14.1.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleads==30.0.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
msgpack==1.0.3
mypy-extensions==0.4.3
nox==2022.1.7
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
proto-plus==1.18.1
protobuf==3.17.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
tomli==2.0.0
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
Werkzeug==2.0.2
xmltodict==0.12.0
zeep==4.1.0
